Disable PIP and fullscreen in iOS PWA. See the issue here: https://anton.codes/ios-pwa-video-issues/
It's not tested on iOS 13 and earlier, as the "Add to home screen" feature in these versions doesn't work on BrowserStack. For iOS 14 and later, PIP in PWA never worked. iOS 14 can indeed go into fullscreen mode without the home indicator permanently showing, but the system volume sliders are still displayed. It's not worth the hassle to handle iOS 14 separately.

We will revisit this issue and re-enable the features if fixed in future versions.